[fix]カラム変更

+++ db/migrate/20210401193431_create_rooms.rb
@@ -1,8 +1,8 @@
 class CreateRooms < ActiveRecord::Migration[5.2]
   def change
     create_table :rooms do |t|
-      t.references :user,type: :integer, foreign_key: true
-      t.references :coach,type: :integer, foreign_key: true
+      t.references :user,type: :bigint, foreign_key: true
+      t.references :coach,type: :bigint, foreign_key: true
 
       t.timestamps
     end
